requirements.txt
blinker==1.7.0
click==8.1.7
dominate==2.9.0
flask==3.0.0
Flask-Bootstrap==3.3.7.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pure-python-adb==0.3.0.dev0
python-nmap==0.7.1
visitor==0.1.3
werkzeug==3.0.1
zipp==3.17.0
